Add chash_iterator.
chash_iterator allows you to iterate over the entries in the table.
Previously the way to do this was to use chash_{keys,values,entries},
but these allocated memory proportional to the size of the table. These
three have been removed; they're easy to implement in terms of a
chash_iterator if needed (as in examples/word_count.c). This lets the
caller manage his own memory.